+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/dataframe/action/GetDataFrameTransformsActionRequestTests.java
@@ -4,15 +4,20 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-package org.elasticsearch.xpack.dataframe.action;
+package org.elasticsearch.xpack.core.dataframe.action;
 
+import org.elasticsearch.cluster.metadata.MetaData;
 import org.elasticsearch.common.io.stream.Writeable;
 import org.elasticsearch.test.AbstractWireSerializingTestCase;
-import org.elasticsearch.xpack.core.dataframe.action.DeleteDataFrameTransformAction.Request;
+import org.elasticsearch.xpack.core.dataframe.action.GetDataFrameTransformsAction.Request;
 
-public class DeleteDataFrameTransformActionRequestTests extends AbstractWireSerializingTestCase<Request> {
+public class GetDataFrameTransformsActionRequestTests extends AbstractWireSerializingTestCase<Request> {
+
     @Override
     protected Request createTestInstance() {
+        if (randomBoolean()) {
+            return new Request(MetaData.ALL);
+        }
         return new Request(randomAlphaOfLengthBetween(1, 20));
     }
