+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/dataframe/transform/MockDeprecatedAggregationBuilder.java
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-package org.elasticsearch.xpack.dataframe.transforms;
+package org.elasticsearch.xpack.core.dataframe.transform;
 
 import org.apache.logging.log4j.LogManager;
 import org.elasticsearch.common.io.stream.StreamInput;
