+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/dataframe/transform/pivot/PivotConfigTests.java
@@ -4,14 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-package org.elasticsearch.xpack.dataframe.transforms.pivot;
+package org.elasticsearch.xpack.core.dataframe.transform.pivot;
 
 import org.elasticsearch.common.io.stream.Writeable.Reader;
 import org.elasticsearch.common.xcontent.DeprecationHandler;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.common.xcontent.XContentType;
-import org.elasticsearch.xpack.core.dataframe.transform.pivot.PivotConfig;
-import org.elasticsearch.xpack.dataframe.transforms.AbstractSerializingDataFrameTestCase;
+import org.elasticsearch.xpack.core.dataframe.transform.AbstractSerializingDataFrameTestCase;
 
 import java.io.IOException;
