+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/dataframe/transform/pivot/TermsGroupSourceTests.java
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-package org.elasticsearch.xpack.dataframe.transforms.pivot;
+package org.elasticsearch.xpack.core.dataframe.transform.pivot;
 
 import org.elasticsearch.common.io.stream.Writeable.Reader;
 import org.elasticsearch.common.xcontent.XContentParser;
