+++ x-pack/plugin/data-frame/src/test/java/org/elasticsearch/xpack/dataframe/transforms/pivot/PivotTests.java
@@ -29,6 +29,7 @@ import org.elasticsearch.search.SearchModule;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.test.client.NoOpClient;
 import org.elasticsearch.xpack.core.dataframe.transform.pivot.AggregationConfig;
+import org.elasticsearch.xpack.core.dataframe.transform.pivot.GroupConfigTests;
 import org.elasticsearch.xpack.core.dataframe.transform.pivot.PivotConfig;
 import org.elasticsearch.xpack.dataframe.transforms.pivot.Aggregations.AggregationType;
 import org.junit.After;
